sort using size_t since negatives are not useful
Change the arguments to the sorting algorithms to use size_t,
which is an unsigned number, since negative values are
not useful.

+++ sort/c/sort.c
@@ -2,7 +2,7 @@
 #include "sort.h"
 
 #if BUBBLE_SORT
-void sort(char **lines, int size)
+void sort(char **lines, size_t size)
 {
 	int i;
 	char *tmp;
@@ -24,7 +24,7 @@ void sort(char **lines, int size)
 	}
 }
 #elif INSERTION_SORT
-void sort(char **lines, int size)
+void sort(char **lines, size_t size)
 {
 	unsigned int i, j;
